Removed comment from PlannedEvent.kt
The comment that the group used to describe this dto class really wasn't necessary. The class, while slightly more difficult to understand than the other dto classes, is still simple enough for most people to understand without explanation. Having too many comments is one of the Developers' Seven Deadly Sins, so it's important to know when comments are necessary and when they aren't