Removed some comments and added others
There were two comments in this file that really didn't need to be there. These comments reiterated what was stated a few lines above them, that being that the function was checking for a left or right swipe. Since it was already stated what the function is doing and the methods have .onSwipeRight() and .onSwipeLeft() in them, it wasn't necessary to repeat that when a person viewing the code could easily figure it out themselves. At the same time, some of the functions and classes in the file that should have had comments briefly explaining what they're meant to do. Having too many comments or not enough comments is one of the Developers' Seven Deadly Sins, so it's important that you include comments when appropriate.