Removed comment from GroupEvent.kt
The comment in this file, much like in PlannedEvents.kt, isn't necessary. This class is simple enough that someone who looks at the code would be able to understand it without explanation. Having too many comments is one of the Developers' Seven Deadly Sins, so you'll want to make sure that comments are only included when necessary.